@@ -1,81 +1,50 @@
-import React from 'react'
-import { makeStyles } from '@material-ui/core/styles';
-import Typography from '@material-ui/core/Typography';
-import pressure from '../assets/pressure.svg'
-import wind_speed from '../assets/wind_speed.svg'
-import humidity from '../assets/humidity.svg'
-import sunrise from '../assets/sunrise.svg'
-import sunset from '../assets/sunset.svg'
-import CardContent from '@material-ui/core/CardContent';
+import React from "react";
 
+import pressure from "../assets/pressure.svg";
+import wind_speed from "../assets/wind_speed.svg";
+import humidity from "../assets/humidity.svg";
+import sunrise from "../assets/sunrise.svg";
+import sunset from "../assets/sunset.svg";
+import { Card, Col, Row, Container } from "react-bootstrap";
+import Font from 'react-font'
+function Today({ today }) {
+  return (
+    <>
+       <Font family='Ubuntu'>
+      <Row className="border bg-light p-2">
+        <Col className="fs-2 fw-bolder bg-info  text-light">{today.temp}°C</Col>
+        <Col className=" bg-primary text-light">
+          {today.main}, {today.desc}
+        </Col>
+        <Col className="fs-4 fw-bolder">
+          {" "}
+          <img src={sunrise} alt="Logo" className="img-fluid col-6" />{" "}
+          {today.sunrise} A.M{" "}
+        </Col>
+        <Col className="fs-4 fw-bolder">
+          <img src={sunset} alt="Logo" className="img-fluid col-6" />{" "}
+          {today.sunset} P.M
+        </Col>
 
-const useStyles = makeStyles((theme) => ({
-    unit__icon: {
-        width: 22,
-        height:22,
-        alignSelf: 'center',
-        marginRight: 4,
-        marginLeft: 20,
-    },
-    unit__icon1: {
-        width: 22,
-        height:22,
-        alignSelf: 'center',
-        fontSize: '15',
-    },
-    weather__icon: {
-        width: 90,
-        height: 90,
-        Top: 0,
-    },
-    main : {
-        overflow: 'auto',
-        padding: 5,
-    },
-    text__left: {
-        float: 'left',
-    },
-    text__right: {
-        float: 'right',
-    },
-    span: {
-        fontWeight: 'bold',
-    }
-
-  }));
-
-function Today({today}) {
-    const classes = useStyles();
-    return (
-        <CardContent>
-            <div className={classes.main}>
-                <div className={classes.text__left}>
-                    <img src={`https://openweathermap.org/img/w/${today.icon}.png`} alt={today.icon} className={classes.weather__icon}/>
-                    <Typography variant="h3" gutterBottom >
-                        {today.temp}°C
-                    </Typography>
-                    <Typography variant="h6" gutterBottom>
-                        {today.main}, {today.desc}
-                    </Typography>
-                </div>
-                <div className={classes.text__right}>
-                    
-                    <Typography variant="h6"gutterBottom >
-                       <img src={sunrise} alt="Logo" className={classes.unit__icon}/> {today.sunrise} A.M.
-                    </Typography>
-                    <Typography variant="h6"gutterBottom >
-                        <img src={sunset} alt="Logo" className={classes.unit__icon}/> {today.sunset} P.M.
-                    </Typography>
-                    
-                </div>
-            </div>
-                <div>
-                    <img src={pressure} alt="Logo" className={classes.unit__icon1}/><span className={classes.span}>{today.pressure} hPa</span>
-                    <img src={humidity} alt="Logo" className={classes.unit__icon}/><span className={classes.span}>{today.humidity} %</span>
-                    <img src={wind_speed} alt="Logo" className={classes.unit__icon}/><span className={classes.span}>{today.wind} m/s N</span>
-                </div>
-            </CardContent>    
-    )
+       
+      </Row>
+      <Row className="mt-1 border p-5 bg-light ">
+      <Col>
+          <img src={pressure} alt="Logo" className="img-fluid col-4 " />
+          <span className="fs-2 fw-bolder">{today.pressure} hPa</span>
+        </Col>
+        <Col>  
+          <img src={humidity} alt="Logo" className="img-fluid col-4 " xs={12} />
+          <span className="fs-2 fw-bolder">{today.humidity} %</span>
+        </Col>
+        <Col>
+          <img src={wind_speed} alt="Logo" className="img-fluid col-4" />
+          <span className="fs-2 fw-bolder">{today.wind} m/s N</span>
+        </Col>
+      </Row>
+      </Font>
+    </>
+  );
 }
 
-export default Today
+export default Today;
